+++ server_extensions_extra/src/main/java/com/generalbytes/batm/server/extensions/extra/groestlcoin/fr/cryptohash/DigestEngine.java
@@ -1,6 +1,6 @@
 // $Id: DigestEngine.java 229 2010-06-16 20:22:27Z tp $
 
-package fr.cryptohash;
+package com.generalbytes.batm.server.extensions.extra.groestlcoin.fr.cryptohash;
 
 /**
  * <p>This class is a template which can be used to implement hash
@@ -20,7 +20,7 @@ package fr.cryptohash;
  * ==========================(LICENSE BEGIN)============================
  *
  * Copyright (c) 2007-2010  Projet RNRT SAPHIR
- * 
+ *
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -28,10 +28,10 @@ package fr.cryptohash;
  * distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to
  * the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
