+++ server_extensions_extra/src/main/java/com/generalbytes/batm/server/extensions/extra/groestlcoin/fr/cryptohash/Groestl512.java
@@ -1,6 +1,6 @@
 // $Id: Groestl512.java 198 2010-05-27 16:45:04Z tp $
 
-package fr.cryptohash;
+package com.generalbytes.batm.server.extensions.extra.groestlcoin.fr.cryptohash;
 
 /**
  * <p>This class implements the Groestl-512 digest algorithm under the
@@ -10,7 +10,7 @@ package fr.cryptohash;
  * ==========================(LICENSE BEGIN)============================
  *
  * Copyright (c) 2007-2010  Projet RNRT SAPHIR
- * 
+ *
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -18,10 +18,10 @@ package fr.cryptohash;
  * distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to
  * the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
